Remove MarketBook loading from json
Loading from json requires an import of MarketDefinition but
this causes import conflicts. This could be fixed by moving
MarketDefinition to a new module but I want to keep this
fork as close as possible to the original repo so I have
just removed the json loading. The loading will how have to
be performed manually where required.